+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Fix serializing nil facts
   - Fix newlines in fact values
   - Don't pre-populate single-choice tuples in `InstanceDocument::from_taxonomy`
+  - Allow missing identifier in `xbrli:entity` during parsing
 - added
   - Make `ArcroleUri`, `ArcroleType`, `RoleType`, `XbrlType`, `Concept`, `Particle`, `ElementDecl`,
     `ElementParticle`, `GroupDef`, `GroupParticle`, `Occurrence`, `ROLE_TERSE`,